+++ dart/lib/src/sentry_tracer.dart
@@ -17,7 +17,9 @@ class SentryTracer extends ISentrySpan {
   late final SentrySpan _rootSpan;
   final List<SentrySpan> _children = [];
   final Map<String, dynamic> _extra = {};
+
   final Map<String, SentryMeasurement> _measurements = {};
+  Map<String, SentryMeasurement> get measurements => _measurements;
 
   Timer? _autoFinishAfterTimer;
   Duration? _autoFinishAfter;
@@ -320,10 +322,6 @@ class SentryTracer extends ISentrySpan {
   @override
   SentryTraceHeader toSentryTrace() => _rootSpan.toSentryTrace();
 
-  @visibleForTesting
-  Map<String, SentryMeasurement> get measurements =>
-      Map.unmodifiable(_measurements);
-
   bool _haveAllChildrenFinished() {
     for (final child in children) {
       if (!child.finished) {
@@ -341,10 +339,19 @@ class SentryTracer extends ISentrySpan {
   @override
   void setMeasurement(String name, num value, {SentryMeasurementUnit? unit}) {
     if (finished) {
+      _hub.options.logger(SentryLevel.debug,
+          "The tracer is already finished. Measurement $name cannot be set");
       return;
     }
-    final measurement = SentryMeasurement(name, value, unit: unit);
-    _measurements[name] = measurement;
+    _measurements[name] = SentryMeasurement(name, value, unit: unit);
+  }
+
+  void setMeasurementFromChild(String name, num value,
+      {SentryMeasurementUnit? unit}) {
+    // We don't want to overwrite span measurement, if it comes from a child.
+    if (!_measurements.containsKey(name)) {
+      setMeasurement(name, value, unit: unit);
+    }
   }
 
   @override